requirements.txt
Django>=4.0
djangorestframework>=3.14
kafka-python>=2.0
psycopg2>=2.9
redis>=4.0
